requirements.txt
fastapi
uvicorn
pymongo
google-genai
pydantic
reportlab
python-dotenv